Move fifo_cache to pkg/chunk/cache
