+++ pkg/chunk/cache/fifo_cache_test.go
@@ -1,4 +1,4 @@
-package chunk
+package cache
 
 import (
 	"context"
